[OPR-54]: Completes the refactor of previous functionality.
